+++ ChangeLog.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - added REPL multi-line support
 
+### fixed
+
+- fixed handling of EOF in triple quotes string 
+
 
 
 ## [1.3.4] - 2019-02-22